EA-2193 Move unit tests to src/tests/unit/ following repo convention
Main moved tests from colocated src/dataLayer/*.test.js to
src/tests/unit/dataLayer/. Move all new unit tests to match:

- schemaRegistry.test.js -> src/tests/unit/dataLayer/clickHouse/
- genericClickHouseQueryParser.test.js -> src/tests/unit/dataLayer/clickHouse/
- genericClickHouseQueryBuilder.test.js -> src/tests/unit/dataLayer/builders/
- genericClickHouseRepository.test.js -> src/tests/unit/dataLayer/repositories/
- clickHouseDatabaseCursor.test.js -> src/tests/unit/dataLayer/databaseCursor/
- clickHouseBulkWriteExecutor.test.js -> src/tests/unit/dataLayer/bulkWriteExecutors/
- executorRouting.test.js -> src/tests/unit/dataLayer/bulkWriteExecutors/

Import paths updated. Integration tests remain in src/tests/clickhouseOnly/.
